Added BBQ charcoal mice
Recipe: Press Space to put down the mouse BBQ device :bomb:; wait 100 ticks for it to activate; mice within a radius of 1.5 tiles will be instantly barbequed into charcoal. Press O key to order more mice. Caution: Do NOT stand in a raduis of 1.5 tiles when the BBQ device activate. And some refactoring, and made the mouse a little bit smaller.